requirements.txt
Flask==3.0.3
qrcode==7.4.2
Pillow==10.3.0